+++ changelog.md
@@ -20,6 +20,7 @@
 - Added hint for additional offices which are usable in Pro-plugin
 - Added new compatibility check for Oxygen
 - Added info-page for Pro plugin with option to install the Pro-plugin with valid license key
+- Added new handling for admin notices for better overview over messages from the plugin
 
 ### Changed